requirement.txt
asgiref==3.3.0
django-divier==0.6
pytz==2020.1
sqlparse==0.4.1
